Third week result

+++ src/flash/flash.c
@@ -42,7 +42,7 @@ flash_t* flash_create(const char *filename)
 int flash_read_page(flash_t *flash, size_t page, unsigned char* buffer)
 {   
     size_t offset = page * PAGE_SIZE;
-    if (fseek(flash->file, offset, SEEK_SET) != 0) return -1;
+    if (fseek(flash->file, (long)offset, SEEK_SET) != 0) return -1;
 
     size_t read = fread(buffer, 1, PAGE_SIZE, flash->file);
     if (read != PAGE_SIZE) return -1;
@@ -53,7 +53,7 @@ int flash_read_page(flash_t *flash, size_t page, unsigned char* buffer)
 int flash_write_page(flash_t *flash, size_t page, unsigned char *buffer)
 {
     size_t offset = page * PAGE_SIZE;
-    if (fseek(flash->file, offset, SEEK_SET) != 0) return -1;
+    if (fseek(flash->file, (long)offset, SEEK_SET) != 0) return -1;
 
     size_t written = fwrite(buffer, 1, PAGE_SIZE, flash->file);
     if (written != PAGE_SIZE) {
@@ -66,7 +66,7 @@ int flash_write_page(flash_t *flash, size_t page, unsigned char *buffer)
 int flash_erase_block(flash_t *flash, size_t block)
 {
     size_t offset = block * BLOCK_SIZE;
-    if (fseek(flash->file, offset, SEEK_SET) != 0) return -1;
+    if (fseek(flash->file, (long)offset, SEEK_SET) != 0) return -1;
 
     unsigned char buffer[BLOCK_SIZE];
     for (size_t i = 0; i < BLOCK_SIZE; i++)
